split GroundItemInputListener into GroundItemMouseAdapter and GroundItemHotkeyListener
