+++ runelite-client/src/main/java/net/runelite/client/util/HotkeyListener.java
@@ -85,10 +85,15 @@ public abstract class HotkeyListener implements KeyListener
 		{
 			isPressed = false;
 			isConsumingTyped = false;
+			hotkeyReleased();
 		}
 	}
 
 	public void hotkeyPressed()
+	{
+	}
+
+	public void hotkeyReleased()
 	{
 	}
 }
